@@ -11,24 +11,25 @@
         }
     ],
     "require": {
-        "php": "^7.2 || ^8.0",
+        "php": "^8.0.2",
         "ext-json": "*",
-        "composer/composer": "^1.0 || ^2.0",
-        "illuminate/console": "^6.0 || ^7.0 || ^8.0 || ^9.0",
-        "illuminate/container": "^6.0 || ^7.0 || ^8.0 || ^9.0",
-        "illuminate/contracts": "^6.0 || ^7.0 || ^8.0 || ^9.0",
-        "illuminate/database": "^6.0 || ^7.0 || ^8.0 || ^9.0",
-        "illuminate/http": "^6.0 || ^7.0 || ^8.0 || ^9.0",
-        "illuminate/pipeline": "^6.0 || ^7.0 || ^8.0 || ^9.0",
-        "illuminate/support": "^6.0 || ^7.0 || ^8.0 || ^9.0",
-        "mockery/mockery": "^0.9 || ^1.0",
-        "phpstan/phpstan": "^1.0",
-        "symfony/process": "^4.3 || ^5.0 || ^6.0"
+        "composer/class-map-generator": "^1.0",
+        "composer/pcre": "^3.0",
+        "illuminate/console": "^9",
+        "illuminate/container": "^9",
+        "illuminate/contracts": "^9",
+        "illuminate/database": "^9",
+        "illuminate/http": "^9",
+        "illuminate/pipeline": "^9",
+        "illuminate/support": "^9",
+        "mockery/mockery": "^1.4.4",
+        "phpmyadmin/sql-parser": "^5.5",
+        "phpstan/phpstan": "^1.8.1"
     },
     "require-dev": {
-        "nikic/php-parser": "^4.13.0",
-        "orchestra/testbench": "^4.0 || ^5.0 || ^6.0 || ^7.0",
-        "phpunit/phpunit": "^7.3 || ^8.2 || ^9.3"
+        "nikic/php-parser": "^4.13.2",
+        "orchestra/testbench": "^7.0.0",
+        "phpunit/phpunit": "^9.5.11"
     },
     "suggest": {
         "orchestra/testbench": "Using Larastan for analysing a package needs Testbench"
@@ -36,9 +37,8 @@
     "autoload-dev": {
         "psr-4": {
             "App\\": "tests/Application/app/",
-            "Laravel8\\": "tests/Laravel8Application/app/",
-            "Database\\Factories\\": "tests/Laravel8Application/database/factories/",
-            "Database\\Migrations\\": "tests/Laravel8Application/database/migrations/"
+            "Database\\Factories\\": "tests/Application/database/factories/",
+            "Database\\Migrations\\": "tests/Application/database/migrations/"
         },
       "classmap": [
         "tests/"
@@ -57,7 +57,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.0-dev"
+            "dev-master": "2.0-dev"
         },
         "phpstan": {
             "includes": [
@@ -66,8 +66,8 @@
         }
     },
     "scripts": {
-        "test:types": "phpstan analyse --ansi",
-        "test:unit": "phpunit --colors=always",
+        "test:types": "phpstan analyse --ansi --memory-limit 256M",
+        "test:unit": "phpunit --colors=always -d memory_limit=512M",
         "test": [
             "@test:types",
             "@test:unit"
