@@ -14,7 +14,13 @@ use PhpParser\NodeFinder;
 final class SchemaAggregator
 {
     /** @var array<string, SchemaTable> */
-    public $tables = [];
+    public array $tables = [];
+
+    /** @param array<string, SchemaTable> $tables */
+    public function __construct(array $tables = [])
+    {
+        $this->tables = $tables;
+    }
 
     /**
      * @param  array<int, PhpParser\Node\Stmt>  $stmts
@@ -56,33 +62,50 @@ final class SchemaAggregator
 
         foreach ($methods as $stmt) {
             if ($stmt instanceof PhpParser\Node\Stmt\Expression
+                && $stmt->expr instanceof PhpParser\Node\Expr\MethodCall
+                && $stmt->expr->var instanceof PhpParser\Node\Expr\StaticCall
+                && $stmt->expr->var->class instanceof PhpParser\Node\Name
+                && $stmt->expr->var->name instanceof PhpParser\Node\Identifier
+                && ($stmt->expr->var->name->toString() === 'connection' || $stmt->expr->var->name->toString() === 'setConnection')
+                && ($stmt->expr->var->class->toCodeString() === '\Illuminate\Support\Facades\Schema' || $stmt->expr->var->class->toCodeString() === '\Schema')
+            ) {
+                $statement = $stmt->expr;
+            } elseif ($stmt instanceof PhpParser\Node\Stmt\Expression
                 && $stmt->expr instanceof PhpParser\Node\Expr\StaticCall
-                && ($stmt->expr->class instanceof PhpParser\Node\Name)
+                && $stmt->expr->class instanceof PhpParser\Node\Name
                 && $stmt->expr->name instanceof PhpParser\Node\Identifier
                 && ($stmt->expr->class->toCodeString() === '\Illuminate\Support\Facades\Schema' || $stmt->expr->class->toCodeString() === '\Schema')
             ) {
-                switch ($stmt->expr->name->name) {
-                    case 'create':
-                        $this->alterTable($stmt->expr, true);
-                        break;
+                $statement = $stmt->expr;
+            } else {
+                continue;
+            }
 
-                    case 'table':
-                        $this->alterTable($stmt->expr, false);
-                        break;
+            if (! $statement->name instanceof PhpParser\Node\Identifier) {
+                continue;
+            }
 
-                    case 'drop':
-                    case 'dropIfExists':
-                        $this->dropTable($stmt->expr);
-                        break;
+            switch ($statement->name->name) {
+                case 'create':
+                    $this->alterTable($statement, true);
+                    break;
 
-                    case 'rename':
-                        $this->renameTable($stmt->expr);
-                }
+                case 'table':
+                    $this->alterTable($statement, false);
+                    break;
+
+                case 'drop':
+                case 'dropIfExists':
+                    $this->dropTable($statement);
+                    break;
+
+                case 'rename':
+                    $this->renameTableThroughStaticCall($statement);
             }
         }
     }
 
-    private function alterTable(PhpParser\Node\Expr\StaticCall $call, bool $creating): void
+    private function alterTable(PhpParser\Node\Expr\StaticCall|PhpParser\Node\Expr\MethodCall $call, bool $creating): void
     {
         if (! isset($call->args[0])
             || ! $call->getArgs()[0]->value instanceof PhpParser\Node\Scalar\String_
@@ -257,6 +280,7 @@ final class SchemaAggregator
                         case 'unsignedsmallinteger':
                         case 'unsignedtinyinteger':
                         case 'bigincrements':
+                        case 'foreignid':
                             $table->setColumn(new SchemaColumn($columnName, 'int', $nullable));
                             break;
 
@@ -359,6 +383,9 @@ final class SchemaAggregator
                             break;
 
                         case 'rename':
+                            $this->renameTableThroughMethodCall($table, $stmt->expr);
+                            break;
+
                         case 'renamecolumn':
                             if ($secondArg instanceof PhpParser\Node\Scalar\String_) {
                                 $table->renameColumn($columnName, $secondArg->value);
@@ -392,7 +419,7 @@ final class SchemaAggregator
         }
     }
 
-    private function dropTable(PhpParser\Node\Expr\StaticCall $call): void
+    private function dropTable(PhpParser\Node\Expr\StaticCall|PhpParser\Node\Expr\MethodCall $call): void
     {
         if (! isset($call->args[0])
             || ! $call->getArgs()[0]->value instanceof PhpParser\Node\Scalar\String_
@@ -405,7 +432,7 @@ final class SchemaAggregator
         unset($this->tables[$tableName]);
     }
 
-    private function renameTable(PhpParser\Node\Expr\StaticCall $call): void
+    private function renameTableThroughStaticCall(PhpParser\Node\Expr\StaticCall|PhpParser\Node\Expr\MethodCall $call): void
     {
         if (! isset($call->args[0], $call->args[1])
             || ! $call->getArgs()[0]->value instanceof PhpParser\Node\Scalar\String_
@@ -417,6 +444,28 @@ final class SchemaAggregator
         $oldTableName = $call->getArgs()[0]->value->value;
         $newTableName = $call->getArgs()[1]->value->value;
 
+        $this->renameTable($oldTableName, $newTableName);
+    }
+
+    private function renameTableThroughMethodCall(SchemaTable $oldTable, PhpParser\Node\Expr\MethodCall $call): void
+    {
+        if (! isset($call->args[0])
+            || ! $call->getArgs()[0]->value instanceof PhpParser\Node\Scalar\String_
+        ) {
+            return;
+        }
+
+        /** @var PhpParser\Node\Scalar\String_ $methodCallArgument */
+        $methodCallArgument = $call->getArgs()[0]->value;
+
+        $oldTableName = $oldTable->name;
+        $newTableName = $methodCallArgument->value;
+
+        $this->renameTable($oldTableName, $newTableName);
+    }
+
+    private function renameTable(string $oldTableName, string $newTableName): void
+    {
         if (! isset($this->tables[$oldTableName])) {
             return;
         }
