@@ -7,6 +7,7 @@ namespace NunoMaduro\Larastan\Reflection;
 use Illuminate\Database\Eloquent\Builder;
 use PHPStan\Reflection\ClassMemberReflection;
 use PHPStan\Reflection\ClassReflection;
+use PHPStan\Reflection\ExtendedMethodReflection;
 use PHPStan\Reflection\FunctionVariant;
 use PHPStan\Reflection\MethodReflection;
 use PHPStan\Reflection\ParameterReflection;
@@ -15,7 +16,8 @@ use PHPStan\Type\Generic\TemplateTypeMap;
 use PHPStan\Type\ObjectType;
 use PHPStan\Type\Type;
 
-final class EloquentBuilderMethodReflection implements MethodReflection
+/** @phpstan-ignore-next-line */
+final class EloquentBuilderMethodReflection implements ExtendedMethodReflection
 {
     /**
      * @var string
