@@ -4,15 +4,23 @@ declare(strict_types=1);
 
 namespace NunoMaduro\Larastan\ReturnTypes;
 
+use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Database\Eloquent\Model;
+use NunoMaduro\Larastan\Types\Factory\ModelFactoryType;
 use PhpParser\Node\Expr\StaticCall;
 use PhpParser\Node\Name;
 use PHPStan\Analyser\Scope;
 use PHPStan\Reflection\MethodReflection;
+use PHPStan\TrinaryLogic;
+use PHPStan\Type\Accessory\AccessoryNumericStringType;
 use PHPStan\Type\DynamicStaticMethodReturnTypeExtension;
 use PHPStan\Type\ErrorType;
-use PHPStan\Type\ObjectType;
+use PHPStan\Type\FloatType;
+use PHPStan\Type\IntegerType;
+use PHPStan\Type\IntersectionType;
+use PHPStan\Type\StringType;
 use PHPStan\Type\Type;
+use PHPStan\Type\UnionType;
 
 final class ModelFactoryDynamicStaticMethodReturnTypeExtension implements DynamicStaticMethodReturnTypeExtension
 {
@@ -23,16 +31,7 @@ final class ModelFactoryDynamicStaticMethodReturnTypeExtension implements Dynami
 
     public function isStaticMethodSupported(MethodReflection $methodReflection): bool
     {
-        if ($methodReflection->getName() !== 'factory') {
-            return false;
-        }
-
-        // Class only available on Laravel 8
-        if (! class_exists('\Illuminate\Database\Eloquent\Factories\Factory')) {
-            return false;
-        }
-
-        return true;
+        return $methodReflection->getName() === 'factory';
     }
 
     public function getTypeFromStaticMethodCall(
@@ -46,12 +45,35 @@ final class ModelFactoryDynamicStaticMethodReturnTypeExtension implements Dynami
             return new ErrorType();
         }
 
+        if (count($methodCall->getArgs()) === 0) {
+            $isSingleModel = TrinaryLogic::createYes();
+        } else {
+            $argType = $scope->getType($methodCall->getArgs()[0]->value);
+
+            $numericTypes = [
+                new IntegerType(),
+                new FloatType(),
+                new IntersectionType([
+                    new StringType(),
+                    new AccessoryNumericStringType(),
+                ]),
+            ];
+
+            $isSingleModel = (new UnionType($numericTypes))->isSuperTypeOf($argType)->negate();
+        }
+
+        $factoryName = Factory::resolveFactoryName(ltrim($class->toCodeString(), '\\')); // @phpstan-ignore-line
+
+        if (class_exists($factoryName)) {
+            return new ModelFactoryType($factoryName, null, null, $isSingleModel);
+        }
+
         $modelName = basename(str_replace('\\', '/', $class->toCodeString()));
 
         if (! class_exists('Database\\Factories\\'.$modelName.'Factory')) {
             return new ErrorType();
         }
 
-        return new ObjectType('Database\\Factories\\'.$modelName.'Factory');
+        return new ModelFactoryType('Database\\Factories\\'.$modelName.'Factory', null, null, $isSingleModel);
     }
 }
