@@ -18,10 +18,10 @@ final class ModelRuleHelper
 {
     public function findModelReflectionFromType(Type $type): ?ClassReflection
     {
-        if ((new ObjectType(Builder::class))->isSuperTypeOf($type)->no() &&
-            (new ObjectType(EloquentBuilder::class))->isSuperTypeOf($type)->no() &&
-            (new ObjectType(Relation::class))->isSuperTypeOf($type)->no() &&
-            (new ObjectType(Model::class))->isSuperTypeOf($type)->no()
+        if (! (new ObjectType(Builder::class))->isSuperTypeOf($type)->yes() &&
+            ! (new ObjectType(EloquentBuilder::class))->isSuperTypeOf($type)->yes() &&
+            ! (new ObjectType(Relation::class))->isSuperTypeOf($type)->yes() &&
+            ! (new ObjectType(Model::class))->isSuperTypeOf($type)->yes()
         ) {
             return null;
         }
