@@ -5,8 +5,9 @@ namespace App;
 use Illuminate\Database\Eloquent\Collection;
 
 /**
- * @template TModel
- * @extends Collection<TModel>
+ * @template TKey of array-key
+ * @template TValue of \Illuminate\Database\Eloquent\Model
+ * @extends Collection<TKey, TValue>
  */
 class RoleCollection extends Collection
 {
