@@ -5,8 +5,9 @@ namespace App;
 use Illuminate\Database\Eloquent\Collection;
 
 /**
+ * @template TKey of array-key
  * @template TModel
- * @extends Collection<TModel>
+ * @extends Collection<TKey, TModel>
  */
 class TransactionCollection extends Collection
 {
