@@ -2,6 +2,8 @@
 
 declare(strict_types=1);
 
+namespace Database\Migrations;
+
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 
@@ -12,7 +14,7 @@ class CreateAccountsTable extends Migration
      */
     public function up(): void
     {
-        Schema::create('accounts', static function (Blueprint $table) {
+        \Schema::create('accounts', static function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name');
             $table->string('active', 1);
