@@ -7,10 +7,11 @@ namespace Tests\Features\Properties;
 use App\User;
 use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Collection;
+use function PHPStan\Testing\assertType;
 
 class HigherOrderCollectionProxyProperties
 {
-    /** @var Collection<User> */
+    /** @var Collection<int, User> */
     public $users;
 
     public function testAverage(): float
@@ -23,7 +24,7 @@ class HigherOrderCollectionProxyProperties
         return $this->users->contains->email;
     }
 
-    /** @return Collection<User> */
+    /** @return Collection<int, User> */
     public function testEach(): Collection
     {
         // Does not make too much sense, but it should work
@@ -36,7 +37,7 @@ class HigherOrderCollectionProxyProperties
         return $this->users->every->email;
     }
 
-    /** @return Collection<User> */
+    /** @return Collection<int, User> */
     public function testFilter(): Collection
     {
         // Does not make too much sense, but it should work
@@ -55,18 +56,15 @@ class HigherOrderCollectionProxyProperties
         return $this->users->flatMap->email;
     }
 
-    /** @return Collection<Collection<User>> */
-    public function testGroupBy(): Collection
+    public function testGroupBy(): void
     {
-        // Does not make too much sense, but it should work
-        return $this->users->groupBy->email;
+        assertType('Illuminate\Database\Eloquent\Collection<int, Illuminate\Database\Eloquent\Collection<int, App\User>>', $this->users->groupBy->email);
     }
 
-    /** @return Collection<User> */
-    public function testKeyBy(): Collection
+    public function testKeyBy(): void
     {
         // Does not make too much sense, but it should work
-        return $this->users->keyBy->email;
+        assertType('Illuminate\Database\Eloquent\Collection<(int|string), App\User>', $this->users->keyBy->email);
     }
 
     /** @return \Illuminate\Support\Collection<int, string> */
@@ -120,25 +118,24 @@ class HigherOrderCollectionProxyProperties
         return $this->users->min->created_at;
     }
 
-    /** @return Collection<Collection<User>> */
-    public function testPartition(): Collection
+    public function testPartition(): void
     {
-        return $this->users->partition->email;
+        assertType('Illuminate\Database\Eloquent\Collection<int, Illuminate\Database\Eloquent\Collection<int, App\User>>', $this->users->partition->email);
     }
 
-    /** @return Collection<User> */
+    /** @return Collection<int, User> */
     public function testReject(): Collection
     {
         return $this->users->reject->email;
     }
 
-    /** @return Collection<User> */
+    /** @return Collection<int, User> */
     public function testSkipUntil(): Collection
     {
         return $this->users->skipUntil->email;
     }
 
-    /** @return Collection<User> */
+    /** @return Collection<int, User> */
     public function testSkipWhile(): Collection
     {
         return $this->users->skipWhile->email;
@@ -149,19 +146,19 @@ class HigherOrderCollectionProxyProperties
         return $this->users->sum->id;
     }
 
-    /** @return Collection<User> */
+    /** @return Collection<int, User> */
     public function testTakeUntil(): Collection
     {
         return $this->users->takeUntil->email;
     }
 
-    /** @return Collection<User> */
+    /** @return Collection<int, User> */
     public function testTakeWhile(): Collection
     {
         return $this->users->takeWhile->email;
     }
 
-    /** @return Collection<User> */
+    /** @return Collection<int, User> */
     public function testUnique(): Collection
     {
         return $this->users->unique->email;
