@@ -2,5 +2,5 @@
 
 function foo(): void
 {
-    \Laravel8\Models\User::factory()->createOne(['foo' => 'bar']);
+    \App\User::factory()->createOne(['foo' => 'bar']);
 }
